requirements.txt
google-cloud-secret-manager==2.5.0
TwitterAPI==2.7.3